Modify playlist schema and fix playlist routes
Issue was that upon 'get latest data', all mongoose IDs of albums that were in playlists were no longer valid. Needed to change the structure of how C/D requests were handled to be able to access playlist data regardless of mongoose ID.

Solution was to not 'populate' albums within playlists via mongoose IDs, but to physically store them in there with all of their attributes. This is also advantageous in a situation where enough time has passed that the album is no longer on recent uploads (unlikely but possible), and now playlists are a separate entity from any of the latest artist data.