Call spotify auth from back end using axios
Solves issue (on dev side at least) that was bringing up the CORS error when trying to log into spotify account. Now, the request will be made solely from the backend and just the redirect url is passed to the front end. The front end receives the redirect url and then opens a new window with that link, allowing the front end data to stay unchanged and for the user to log into their account and populate the back end with their access token.

Although not perfect, this is the best implementation thus far of having a complete login flow from the front end that allows the user to access their follows and perform CRUD operations on them.