Fix repeatable sign-up bug, change sign-up flow
User could previously sign up multiple times with the same email and not receive an error (even though email was never repeatedly added to database)

Now the program will first check if the user exists, and if it does not, proceed to check that passwords match. If both criteria pass, the user will be created and passed back to front end.